Breaking the types down into individual files
