+++ src/shared/types/AbilityScores.ts
@@ -1,0 +1,15 @@
+export type AbilityScores =
+  | "strength"
+  | "dexterity"
+  | "constitution"
+  | "intelligence"
+  | "wisdom"
+  | "charisma";
+export const ABILITYSCORES = [
+  "strength",
+  "dexterity",
+  "constitution",
+  "intelligence",
+  "wisdom",
+  "charisma"
+];
