+++ src/shared/types/Mastery.ts
@@ -1,0 +1,2 @@
+export type Mastery = "trained" | "expert" | "master" | "legend";
+export const MASTERY = ["trained", "expert", "master", "legend"];
