+++ src/shared/types/Skills.ts
@@ -1,0 +1,6 @@
+import { Mastery } from "@/shared/types/Mastery";
+
+export class Skill {
+  public name: string = "";
+  public level!: Mastery;
+}
